[rename]: authStore 파일명 useAuthStore로 변경
